backend/pyproject.toml
[project]
name = "backend"
version = "0.1.0"
description = ""
authors = [
    {name = "Azam Afridi",email = "azafridi9547@gmail.com"}
]
requires-python = ">=3.10,<4.0"
dependencies = [
    "django (>=5.1.5,<6.0.0)",
    "djangorestframework (>=3.15.2,<4.0.0)",
    "langchain-community (>=0.3.16,<0.4.0)",
    "dj-rest-auth (>=7.0.1,<8.0.0)",
    "django-allauth (>=65.3.1,<66.0.0)",
    "pyjwt (>=2.10.1,<3.0.0)",
    "cryptography (>=44.0.0,<45.0.0)",
    "openai (>=1.61.0,<2.0.0)",
    "pip (>=25.0,<26.0)",
    "django-cors-headers (>=4.6.0,<5.0.0)",
]


[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"